+++ src/Spinner/Settings/Contracts/Defaults.php
@@ -15,7 +15,7 @@ interface Defaults
     public const DOTS_SUFFIX = '...';
     public const ELLIPSIS = '…';
 
-    public const MAX_FRAMES_COUNT = 60;
+    public const MAX_FRAMES_COUNT = 256;
     public const MAX_FRAME_WIDTH = 10;
 
     public const DEFAULT_INTERVAL = 0.1;